Moved token service to separate project
