--- ArcadiaParties.API/OAuthSettings.cs
@@ -1,13 +1,0 @@
-﻿namespace ArcadiaParties.API
-{
-    public class OAuthSettings
-    {
-        public string ClientId { get; set; }
-
-        public string AuthorizationUrl { get; set; }
-
-        public string TokenUrl { get; set; }
-
-        public string OpenIdConfigurationUrl { get; set; }
-    }
-}
