--- ArcadiaParties.API/Token/ITokenGetterFactory.cs
@@ -1,7 +1,0 @@
-﻿namespace ArcadiaParties.API.Token
-{
-    public interface ITokenGetterFactory
-    {
-        ITokenService CreateTokenGetter(AzureApplication azureApplication);
-    }
-}
