+++ ArcadiaParties.Token/TokenFactory/ITokenServiceFactory.cs
@@ -1,0 +1,7 @@
+﻿namespace ArcadiaParties.Token
+{
+    public interface ITokenServiceFactory
+    {
+        ITokenService CreateTokenService(AzureApplication azureApplication);
+    }
+}
